refactor ovftool validation into vmware driver
