+++ builder/vmware/common/driver_config.go
@@ -3,14 +3,8 @@
 package common
 
 import (
-	"bytes"
-	"context"
 	"fmt"
-	"net/url"
 	"os"
-	"os/exec"
-	"strings"
-	"time"
 
 	"github.com/hashicorp/packer/template/interpolate"
 )
@@ -91,66 +85,6 @@ func (c *DriverConfig) Validate(SkipExport bool) error {
 	if c.RemoteType != "" && c.RemotePassword == "" {
 		return fmt.Errorf("exporting the vm from esxi with ovftool requires " +
 			"that you set a value for remote_password")
-	}
-
-	if c.RemoteType == "" {
-		// Validate that tool exists, but no need to validate credentials.
-		ovftool := GetOVFTool()
-		if ovftool != "" {
-			return nil
-		} else {
-			return fmt.Errorf("Couldn't find ovftool in path! Please either " +
-				"set `skip_export = true` and remove the `format` option " +
-				"from your template, or make sure ovftool is installed on " +
-				"your build system. ")
-		}
-	}
-
-	if c.SkipValidateCredentials {
-		return nil
-	}
-
-	// check that password is valid by sending a dummy ovftool command
-	// now, so that we don't fail for a simple mistake after a long
-	// build
-	ovftool := GetOVFTool()
-
-	// Generate the uri of the host, with embedded credentials
-	ovftool_uri := fmt.Sprintf("vi://%s", c.RemoteHost)
-	u, err := url.Parse(ovftool_uri)
-	if err != nil {
-		return fmt.Errorf("Couldn't generate uri for ovftool: %s", err)
-	}
-	u.User = url.UserPassword(c.RemoteUser, c.RemotePassword)
-
-	ovfToolArgs := []string{"--noSSLVerify", "--verifyOnly", u.String()}
-
-	var out bytes.Buffer
-	cmdCtx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-	defer cancel()
-	cmd := exec.CommandContext(cmdCtx, ovftool, ovfToolArgs...)
-	cmd.Stdout = &out
-
-	// Need to manually close stdin or else the ofvtool call will hang
-	// forever in a situation where the user has provided an invalid
-	// password or username
-	stdin, _ := cmd.StdinPipe()
-	defer stdin.Close()
-
-	if err := cmd.Run(); err != nil {
-		outString := out.String()
-		// The command *should* fail with this error, if it
-		// authenticates properly.
-		if !strings.Contains(outString, "Found wrong kind of object") {
-			err := fmt.Errorf("ovftool validation error: %s; %s",
-				err, outString)
-			if strings.Contains(outString,
-				"Enter login information for source") {
-				err = fmt.Errorf("The username or password you " +
-					"provided to ovftool is invalid.")
-			}
-			return err
-		}
 	}
 
 	return nil
